Modified 002.arithmetical-operations.py

+++ 002.arithmetical-operations.py
@@ -1,10 +1,10 @@
 print(10 + 5)
 print(10 - 5)
 print(10 * 5)
-print(10 / 5)
+print(int(10 / 5)) # int를 붙이지 않으면 소수점이 찍혀서 2.0이 출력됨
 
 import math
 print(math.ceil(2.2)) # 올림
 print(math.floor(2.7)) # 내림
-print(math.pow(2,10)) # 2의 10승. (제곱을 표현할 때 사용)
+print(int(math.pow(2,10))) # 2의 10승. (제곱을 표현할 때 사용)
 print(math.pi) # 파이. 원주율
